+++ src/Models/Queries/PermalinkQuery.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Hirtz\Cms\Tenant\Models\Queries;
+
+use Hirtz\Cms\Tenant\Models\Permalink;
+use Hirtz\Tenant\Models\Queries\Traits\TenantQueryTrait;
+
+/**
+ * @template T of Permalink
+ * @extends \Hirtz\Cms\Models\Queries\PermalinkQuery<T>
+ */
+class PermalinkQuery extends \Hirtz\Cms\Models\Queries\PermalinkQuery
+{
+    use TenantQueryTrait;
+}